Move visualizer test main() into separate file
